+++ apps/api/src/services/data-provider/yahoo-finance/yahoo-finance.service.ts
@@ -138,8 +138,7 @@ export class YahooFinanceService implements DataProviderInterface {
           marketPrice: this.getConvertedValue({
             symbol: aSymbol,
             value: historicalItem.close
-          }),
-          performance: historicalItem.open - historicalItem.close
+          })
         };
       }